CHE-84 Profile controller directory restructure
