+++ server/controllers/profileController/index.ts
@@ -1,0 +1,6 @@
+import createProfile from './createProfile/createProfile';
+import getAllProfiles from './getAllProfiles/getAllProfiles';
+import getProfileById from './getProfileById/getProfileById';
+import updateProfile from './updateProfile/updateProfile';
+
+export { createProfile, getAllProfiles, getProfileById, updateProfile };
